refactor(focus): split the workspace focus code into feature-scoped files
